add tof entropy-encoder spec
