+++ Detectors/TOF/reconstruction/src/TOFReconstructionLinkDef.h
@@ -20,5 +20,6 @@
 #pragma link C++ class o2::tof::ClustererTask + ;
 #pragma link C++ class o2::tof::raw::Encoder + ;
 #pragma link C++ class o2::tof::compressed::Decoder + ;
+#pragma link C++ class o2::tof::CTFCoder + ;
 
 #endif
