+++ Detectors/TOF/workflow/CMakeLists.txt
@@ -17,5 +17,5 @@ o2_add_library(TOFWorkflowUtils
                        src/TOFRawWriterSpec.cxx
                        src/CompressedDecodingTask.cxx
                        src/CompressedInspectorTask.cxx
-               PUBLIC_LINK_LIBRARIES O2::Framework O2::DPLUtils O2::TOFBase O2::DataFormatsTOF
-                                     O2::TOFReconstruction)
+                       src/EntropyEncoderSpec.cxx
+               PUBLIC_LINK_LIBRARIES O2::Framework O2::DPLUtils O2::TOFBase O2::DataFormatsTOF O2::TOFReconstruction)
